added media query for desktop
In mobile version:
Adjusted the card section.
Added a picture as a background in card section.
Changed textcolor on h2 headings and the form.
Adjusted the size of the last grey empty container.

in desktop:
Made a media query to start design of desktop version.
Made navbar visible and created the links and buttons.
Began work with topsection and the heading and picture. Added grid. Work in progress.